refactor(workspace): split module by concern (Phase 3)
Implements Phase 3 of the Rust structure refactor roadmap. Splits the
1551-line `src/workspace/mod.rs` into five submodules, each owning a
single concern, while preserving the existing public API through
re-exports.

Split:

- `paths.rs` (101 LOC) — `home_dir`, `expand_tilde`, `normalize_path`,
  `resolve_path`. Pure path utilities.
- `mounts.rs` (205 LOC) — `parse_mount_spec{,_resolved}`,
  `validate_mount_specs`, `validate_mount_paths`, `validate_mounts`,
  `covers` (crate-visible helper shared with `planner`).
- `planner.rs` (715 LOC, +427) — now also owns `plan_collapse`,
  `CollapsePlan`, `Removal`, `CollapseError`. The create/edit planning
  functions introduced in Phase 2 stay here; the collapse primitive
  joins them so all "planning" logic lives in one module.
- `sensitive.rs` (167 LOC) — `SensitiveMount`, `find_sensitive_mounts`,
  `confirm_sensitive_mounts`.
- `resolve.rs` (488 LOC) — `LoadWorkspaceInput`, `ResolvedWorkspace`,
  `current_dir_workspace`, `saved_workspace_match_depth`,
  `host_path_match_depth`, `resolve_load_workspace`. Everything
  involved in turning a `LoadWorkspaceInput` into a `ResolvedWorkspace`
  at runtime.

Remaining in `mod.rs` (209 LOC, down from 1551):

- The public data types `MountConfig`, `WorkspaceConfig`,
  `WorkspaceEdit` (stable cross-module surface).
- `validate_workspace_config` — composite validator that operates on
  `WorkspaceConfig` holistically.
- The 12 workspace-level validation tests that exercise
  `validate_workspace_config`.
- `pub use` re-exports of every previously-public item, so every
  existing `crate::workspace::X` and `jackin::workspace::X` caller
  continues to resolve without changes.

Verbatim moves. No function body edits. No renames. No logic changes.

Test Preservation Policy honored: all 418 tests preserved, co-located
with the functions they exercise:

- `paths.rs` tests — 7 `resolve_path_*` / `normalize_path_*` tests.
- `mounts.rs` tests — 9 `parse_mount_spec_*` tests.
- `planner.rs` tests — now also holds the 7 `covers_*` tests and 14
  `plan_collapse_*` tests (plus the 9 create/edit tests from Phase 2).
- `resolve.rs` tests — 9 resolution tests including the large
  `resolve_with_ad_hoc_*` and `resolves_saved_workspace_*` cases.
- `mod.rs` tests — 12 remaining `validate_*` tests for
  `validate_workspace_config`.

No test deleted. No test weakened. No assertion dropped.

Call sites outside `src/workspace/`: zero edits required. The
`pub use` re-exports in `mod.rs` preserve every existing import path.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
