+++ src/workspace/mod.rs
@@ -1,7 +1,22 @@
+pub mod mounts;
+pub mod paths;
 pub(crate) mod planner;
+pub mod resolve;
+pub mod sensitive;
+
+pub use mounts::{
+    parse_mount_spec, parse_mount_spec_resolved, validate_mount_paths, validate_mount_specs,
+    validate_mounts,
+};
+pub use paths::{expand_tilde, resolve_path};
+pub use planner::{CollapseError, CollapsePlan, Removal, plan_collapse};
+pub use resolve::{
+    LoadWorkspaceInput, ResolvedWorkspace, current_dir_workspace, resolve_load_workspace,
+    saved_workspace_match_depth,
+};
+pub use sensitive::{SensitiveMount, confirm_sensitive_mounts, find_sensitive_mounts};
 
 use serde::{Deserialize, Serialize};
-use std::path::{Component, Path, PathBuf};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct MountConfig {
@@ -33,334 +48,6 @@ pub struct WorkspaceEdit {
     pub allowed_agents_to_add: Vec<String>,
     pub allowed_agents_to_remove: Vec<String>,
     pub default_agent: Option<Option<String>>,
-}
-
-fn home_dir() -> Option<String> {
-    directories::BaseDirs::new().map(|b| b.home_dir().display().to_string())
-}
-
-pub fn expand_tilde(path: &str) -> String {
-    if (path == "~" || path.starts_with("~/"))
-        && let Some(home) = home_dir()
-    {
-        return path.replacen('~', &home, 1);
-    }
-
-    path.to_string()
-}
-
-/// Normalize an absolute path by resolving `.` and `..` components without
-/// touching the filesystem (unlike [`std::fs::canonicalize`]).
-fn normalize_path(path: &Path) -> PathBuf {
-    let mut parts: Vec<Component<'_>> = Vec::new();
-    for component in path.components() {
-        match component {
-            Component::ParentDir => {
-                if let Some(Component::Normal(_)) = parts.last() {
-                    parts.pop();
-                }
-            }
-            Component::CurDir => {}
-            c => parts.push(c),
-        }
-    }
-    parts.iter().collect()
-}
-
-/// Expand tilde, resolve relative paths to absolute using the current working
-/// directory, and normalize `.` / `..` components.
-pub fn resolve_path(path: &str) -> String {
-    let expanded = expand_tilde(path);
-    let abs = if expanded.starts_with('/') {
-        PathBuf::from(&expanded)
-    } else if let Ok(cwd) = std::env::current_dir() {
-        cwd.join(&expanded)
-    } else {
-        return expanded;
-    };
-    normalize_path(&abs).display().to_string()
-}
-
-pub fn parse_mount_spec(spec: &str) -> anyhow::Result<MountConfig> {
-    Ok(parse_mount_spec_inner(spec, false))
-}
-
-/// Like [`parse_mount_spec`] but also resolves relative paths to absolute.
-/// Use this for CLI arguments where the user may pass relative paths.
-pub fn parse_mount_spec_resolved(spec: &str) -> anyhow::Result<MountConfig> {
-    Ok(parse_mount_spec_inner(spec, true))
-}
-
-fn parse_mount_spec_inner(spec: &str, resolve: bool) -> MountConfig {
-    let (raw, readonly) = spec
-        .strip_suffix(":ro")
-        .map_or((spec, false), |value| (value, true));
-    let (src, dst) = raw
-        .split_once(':')
-        .map_or_else(|| (raw, raw), |(s, d)| (s, d));
-    let expand = if resolve { resolve_path } else { expand_tilde };
-    let expanded_src = expand(src);
-    let dst = if src == dst {
-        expanded_src.clone()
-    } else {
-        dst.to_string()
-    };
-
-    MountConfig {
-        src: expanded_src,
-        dst,
-        readonly,
-    }
-}
-
-/// Structural validation: absolute paths, no duplicate destinations.
-/// Safe to call at config-save time — does not touch the filesystem.
-pub fn validate_mount_specs(mounts: &[MountConfig]) -> anyhow::Result<()> {
-    let mut seen_dst = std::collections::HashSet::new();
-
-    for mount in mounts {
-        if !Path::new(&mount.src).is_absolute() {
-            anyhow::bail!("mount source must be absolute: {}", mount.src);
-        }
-        if !mount.dst.starts_with('/') {
-            anyhow::bail!("mount destination must be an absolute path: {}", mount.dst);
-        }
-        if !seen_dst.insert(mount.dst.clone()) {
-            anyhow::bail!("duplicate mount destination: {}", mount.dst);
-        }
-    }
-
-    Ok(())
-}
-
-/// Filesystem validation: checks that mount sources exist on disk.
-/// Call at load/resolve time, not at config-save time.
-pub fn validate_mount_paths(mounts: &[MountConfig]) -> anyhow::Result<()> {
-    for mount in mounts {
-        if !Path::new(&mount.src).exists() {
-            anyhow::bail!("mount source does not exist: {}", mount.src);
-        }
-    }
-
-    Ok(())
-}
-
-/// Full validation: structural + filesystem checks combined.
-pub fn validate_mounts(mounts: &[MountConfig]) -> anyhow::Result<()> {
-    validate_mount_specs(mounts)?;
-    validate_mount_paths(mounts)
-}
-
-// ── Rule-C covering predicate ───────────────────────────────────────────
-
-/// Returns true iff `parent` strictly covers `child` under rule C:
-/// `parent.src` is a proper ancestor of `child.src`, AND the path suffix
-/// `child.src - parent.src` equals the path suffix `child.dst - parent.dst`.
-///
-/// Equivalently: `child` projects the same host subtree to the same container
-/// location that `parent` would already expose it at.
-///
-/// Identity (equal src and equal dst) returns false — that case is handled by
-/// upsert-by-dst in `edit_workspace`.
-///
-/// The `readonly` flag is ignored here. Readonly mismatches are caught at
-/// `plan_collapse` level, not in the predicate.
-fn covers(parent: &MountConfig, child: &MountConfig) -> bool {
-    let parent_src = parent.src.trim_end_matches('/');
-    let parent_dst = parent.dst.trim_end_matches('/');
-    let child_src = child.src.trim_end_matches('/');
-    let child_dst = child.dst.trim_end_matches('/');
-
-    // Identity is not covering.
-    if parent_src == child_src && parent_dst == child_dst {
-        return false;
-    }
-
-    // child.src must be strictly under parent.src.
-    let Some(src_suffix) = child_src.strip_prefix(parent_src) else {
-        return false;
-    };
-    if !src_suffix.starts_with('/') {
-        return false;
-    }
-
-    // child.dst must be strictly under parent.dst with the same suffix.
-    let Some(dst_suffix) = child_dst.strip_prefix(parent_dst) else {
-        return false;
-    };
-    src_suffix == dst_suffix
-}
-
-/// A proposed mount-set change produced by [`plan_collapse`]. `kept` is the
-/// mount list with all rule-C-redundant entries removed; `removed` describes
-/// each collapse for operator-facing messaging.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CollapsePlan {
-    pub kept: Vec<MountConfig>,
-    pub removed: Vec<Removal>,
-}
-
-/// Records that `child` was collapsed because it is covered by `covered_by`
-/// under rule C (see [`covers`]).
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Removal {
-    pub child: MountConfig,
-    pub covered_by: MountConfig,
-}
-
-/// Conditions that prevent a silent collapse. The operator must resolve these
-/// by hand before the edit can proceed.
-#[derive(Debug, Clone, PartialEq, Eq, thiserror::Error)]
-pub enum CollapseError {
-    #[error(
-        "mount {parent_src} ({parent_mode}) would subsume {child_src} ({child_mode}), \
-         but the readonly flag differs. Match the flag or remove the child first.",
-        parent_src = parent.src,
-        parent_mode = if parent.readonly { "ro" } else { "rw" },
-        child_src = child.src,
-        child_mode = if child.readonly { "ro" } else { "rw" },
-    )]
-    ReadonlyMismatch {
-        parent: MountConfig,
-        child: MountConfig,
-    },
-    #[error(
-        "mount {child_src} is already covered by existing mount {parent_src}. \
-         Nothing to add.",
-        child_src = child.src,
-        parent_src = parent.src,
-    )]
-    ChildUnderExistingParent {
-        parent: MountConfig,
-        child: MountConfig,
-    },
-}
-
-/// Computes a [`CollapsePlan`] for `mounts`.
-///
-/// `new_indexes` identifies which entries in `mounts` originate from the
-/// current operation (upserts for `edit`, all indexes for `create`). Indexes
-/// outside `mounts.len()` are ignored.
-///
-/// Returns `Err` on the first [`CollapseError`] encountered. On success, every
-/// entry in `kept` is pairwise non-covering (see [`covers`]) — i.e., the
-/// returned mount list is rule-C compliant.
-pub fn plan_collapse(
-    mounts: &[MountConfig],
-    new_indexes: &[usize],
-) -> Result<CollapsePlan, CollapseError> {
-    let mut kept = Vec::new();
-    let mut removed = Vec::new();
-
-    for (i, m) in mounts.iter().enumerate() {
-        let parent = mounts
-            .iter()
-            .enumerate()
-            .find(|(j, p)| *j != i && covers(p, m));
-
-        match parent {
-            Some((j, p)) => {
-                if p.readonly != m.readonly {
-                    return Err(CollapseError::ReadonlyMismatch {
-                        parent: p.clone(),
-                        child: m.clone(),
-                    });
-                }
-                let child_is_new = new_indexes.contains(&i);
-                let parent_is_new = new_indexes.contains(&j);
-                if child_is_new && !parent_is_new {
-                    return Err(CollapseError::ChildUnderExistingParent {
-                        parent: p.clone(),
-                        child: m.clone(),
-                    });
-                }
-                removed.push(Removal {
-                    child: m.clone(),
-                    covered_by: p.clone(),
-                });
-            }
-            None => kept.push(m.clone()),
-        }
-    }
-
-    Ok(CollapsePlan { kept, removed })
-}
-
-// ── Sensitive mount detection ────────────────────────────────────────────
-
-/// Path suffixes that indicate sensitive host directories. A mount source is
-/// considered sensitive when its resolved path ends with one of these suffixes
-/// (after tilde expansion).
-const SENSITIVE_SUFFIXES: &[(&str, &str)] = &[
-    ("/.ssh", "SSH keys and configuration"),
-    ("/.aws", "AWS credentials and configuration"),
-    ("/.gnupg", "GPG keys and trust database"),
-    ("/.config/gcloud", "Google Cloud credentials"),
-    ("/.kube", "Kubernetes credentials and configuration"),
-    ("/.docker", "Docker credentials and configuration"),
-];
-
-/// A mount source that matched a sensitive path pattern.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SensitiveMount {
-    pub src: String,
-    pub reason: String,
-}
-
-/// Return any mounts whose source path matches a known sensitive pattern.
-pub fn find_sensitive_mounts(mounts: &[MountConfig]) -> Vec<SensitiveMount> {
-    let mut hits = Vec::new();
-    for mount in mounts {
-        let normalized = mount.src.trim_end_matches('/');
-        for &(suffix, reason) in SENSITIVE_SUFFIXES {
-            if normalized.ends_with(suffix) || normalized == suffix.trim_start_matches('/') {
-                hits.push(SensitiveMount {
-                    src: mount.src.clone(),
-                    reason: reason.to_string(),
-                });
-                break;
-            }
-        }
-    }
-    hits
-}
-
-/// Display a warning for sensitive mounts and ask the operator to confirm.
-/// Returns `Ok(true)` when the operator confirms, `Ok(false)` when they
-/// decline, and `Err` on I/O errors.
-pub fn confirm_sensitive_mounts(sensitive: &[SensitiveMount]) -> anyhow::Result<bool> {
-    use owo_colors::OwoColorize;
-    use std::io::IsTerminal;
-
-    if sensitive.is_empty() {
-        return Ok(true);
-    }
-
-    if !std::io::stdin().is_terminal() {
-        anyhow::bail!(
-            "sensitive mount paths detected but stdin is not a terminal — cannot prompt for confirmation"
-        );
-    }
-
-    eprintln!(
-        "\n{}",
-        "⚠  Sensitive host paths detected in mounts:"
-            .yellow()
-            .bold()
-    );
-    for hit in sensitive {
-        eprintln!("     {} — {}", hit.src.bold(), hit.reason);
-    }
-    eprintln!(
-        "   {}",
-        "These paths may expose credentials to the agent container.".dimmed()
-    );
-    eprintln!();
-
-    Ok(dialoguer::Confirm::new()
-        .with_prompt("Continue with these mounts?")
-        .default(false)
-        .interact()?)
 }
 
 pub fn validate_workspace_config(name: &str, workspace: &WorkspaceConfig) -> anyhow::Result<()> {
@@ -402,622 +89,9 @@ pub fn validate_workspace_config(name: &str, workspace: &WorkspaceConfig) -> any
     Ok(())
 }
 
-pub fn current_dir_workspace(cwd: &Path) -> anyhow::Result<WorkspaceConfig> {
-    let cwd = cwd.canonicalize()?;
-    let path = cwd.display().to_string();
-
-    Ok(WorkspaceConfig {
-        workdir: path.clone(),
-        mounts: vec![MountConfig {
-            src: path.clone(),
-            dst: path,
-            readonly: false,
-        }],
-        allowed_agents: vec![],
-        default_agent: None,
-        last_agent: None,
-    })
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum LoadWorkspaceInput {
-    CurrentDir,
-    Path { src: String, dst: String },
-    Saved(String),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ResolvedWorkspace {
-    pub label: String,
-    pub workdir: String,
-    pub mounts: Vec<MountConfig>,
-}
-
-fn host_path_match_depth(path: &str, canonical_cwd: &Path) -> Option<usize> {
-    let expanded = expand_tilde(path);
-    let canonical_path = Path::new(&expanded).canonicalize().ok()?;
-
-    if canonical_cwd == canonical_path || canonical_cwd.starts_with(&canonical_path) {
-        Some(canonical_path.components().count())
-    } else {
-        None
-    }
-}
-
-pub fn saved_workspace_match_depth(workspace: &WorkspaceConfig, cwd: &Path) -> Option<usize> {
-    let canonical_cwd = cwd.canonicalize().ok()?;
-
-    std::iter::once(host_path_match_depth(&workspace.workdir, &canonical_cwd))
-        .chain(
-            workspace
-                .mounts
-                .iter()
-                .map(|mount| host_path_match_depth(&mount.src, &canonical_cwd)),
-        )
-        .flatten()
-        .max()
-}
-
-pub fn resolve_load_workspace(
-    config: &crate::config::AppConfig,
-    selector: &crate::selector::ClassSelector,
-    cwd: &Path,
-    input: LoadWorkspaceInput,
-    ad_hoc_mounts: &[MountConfig],
-) -> anyhow::Result<ResolvedWorkspace> {
-    let (mut workspace, label) = match input {
-        LoadWorkspaceInput::CurrentDir => {
-            let ws = current_dir_workspace(cwd)?;
-            let label = ws.workdir.clone();
-            (ws, label)
-        }
-        LoadWorkspaceInput::Path { src, dst } => {
-            let expanded_src = expand_tilde(&src);
-            let abs_src = if Path::new(&expanded_src).is_absolute() {
-                PathBuf::from(&expanded_src)
-            } else {
-                cwd.join(&expanded_src)
-            };
-            let canonical_src = abs_src
-                .canonicalize()
-                .map_err(|e| anyhow::anyhow!("cannot resolve path {expanded_src}: {e}"))?;
-            let src_str = canonical_src.display().to_string();
-            let workdir = if dst == src || dst == expanded_src {
-                src_str.clone()
-            } else {
-                dst.clone()
-            };
-            let ws = WorkspaceConfig {
-                workdir,
-                mounts: vec![MountConfig {
-                    src: src_str,
-                    dst: if dst == src || dst == expanded_src {
-                        canonical_src.display().to_string()
-                    } else {
-                        dst
-                    },
-                    readonly: false,
-                }],
-                allowed_agents: vec![],
-                default_agent: None,
-                last_agent: None,
-            };
-            let label = ws.workdir.clone();
-            (ws, label)
-        }
-        LoadWorkspaceInput::Saved(name) => {
-            let workspace = config
-                .workspaces
-                .get(&name)
-                .ok_or_else(|| anyhow::anyhow!("unknown workspace {name}"))?
-                .clone();
-            if !workspace.allowed_agents.is_empty()
-                && !workspace
-                    .allowed_agents
-                    .iter()
-                    .any(|agent| agent == &selector.key())
-            {
-                anyhow::bail!(
-                    "agent {} is not allowed by workspace {name}",
-                    selector.key()
-                );
-            }
-            (workspace, name)
-        }
-    };
-
-    // Merge ad-hoc mounts after workspace mounts, checking for dst conflicts.
-    for ad_hoc in ad_hoc_mounts {
-        if workspace
-            .mounts
-            .iter()
-            .any(|existing| existing.dst == ad_hoc.dst)
-        {
-            anyhow::bail!(
-                "ad-hoc mount destination conflicts with workspace mount destination: {}",
-                ad_hoc.dst
-            );
-        }
-        workspace.mounts.push(ad_hoc.clone());
-    }
-
-    validate_workspace_config("runtime", &workspace)?;
-    validate_mount_paths(&workspace.mounts)?;
-
-    let mut mounts = workspace.mounts.clone();
-    let global_mounts = crate::config::AppConfig::expand_and_validate_named_mounts(
-        &config.resolve_mounts(selector),
-    )?;
-
-    for mount in global_mounts {
-        if mounts.iter().any(|existing| existing.dst == mount.dst) {
-            anyhow::bail!(
-                "global mount destination conflicts with workspace destination: {}",
-                mount.dst
-            );
-        }
-        mounts.push(mount);
-    }
-
-    Ok(ResolvedWorkspace {
-        label,
-        workdir: workspace.workdir,
-        mounts,
-    })
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tempfile::tempdir;
-
-    #[test]
-    fn parses_mount_spec_with_optional_readonly_suffix() {
-        let mount = parse_mount_spec("/tmp/cache:/workspace/cache:ro").unwrap();
-
-        assert_eq!(mount.src, "/tmp/cache");
-        assert_eq!(mount.dst, "/workspace/cache");
-        assert!(mount.readonly);
-    }
-
-    #[test]
-    fn parses_mount_spec_with_src_only() {
-        let mount = parse_mount_spec("/tmp/project").unwrap();
-
-        assert_eq!(mount.src, "/tmp/project");
-        assert_eq!(mount.dst, "/tmp/project");
-        assert!(!mount.readonly);
-    }
-
-    #[test]
-    fn parses_mount_spec_with_src_only_readonly() {
-        let mount = parse_mount_spec("/tmp/project:ro").unwrap();
-
-        assert_eq!(mount.src, "/tmp/project");
-        assert_eq!(mount.dst, "/tmp/project");
-        assert!(mount.readonly);
-    }
-
-    #[test]
-    fn parses_mount_spec_with_tilde_src_only() {
-        let home = std::env::var("HOME").unwrap();
-        let mount = parse_mount_spec("~/projects").unwrap();
-
-        assert_eq!(mount.src, format!("{home}/projects"));
-        assert_eq!(mount.dst, format!("{home}/projects"));
-        assert!(!mount.readonly);
-    }
-
-    #[test]
-    fn resolve_path_resolves_relative_to_cwd() {
-        let cwd = std::env::current_dir().unwrap();
-        let resolved = resolve_path("my-project");
-
-        assert_eq!(resolved, cwd.join("my-project").display().to_string());
-        assert!(resolved.starts_with('/'));
-    }
-
-    #[test]
-    fn resolve_path_leaves_absolute_unchanged() {
-        assert_eq!(resolve_path("/workspace/project"), "/workspace/project");
-    }
-
-    #[test]
-    fn resolve_path_normalizes_dot_to_cwd() {
-        let cwd = std::env::current_dir().unwrap();
-        let resolved = resolve_path(".");
-
-        assert_eq!(resolved, cwd.display().to_string());
-    }
-
-    #[test]
-    fn resolve_path_normalizes_parent_component() {
-        let cwd = std::env::current_dir().unwrap();
-        let resolved = resolve_path("../sibling");
-        let expected = cwd.parent().unwrap().join("sibling");
-
-        assert_eq!(resolved, expected.display().to_string());
-        assert!(!resolved.contains(".."));
-    }
-
-    #[test]
-    fn resolve_path_normalizes_absolute_with_dotdot() {
-        assert_eq!(resolve_path("/a/b/../c"), "/a/c");
-    }
-
-    #[test]
-    fn normalize_path_handles_multiple_parent_refs() {
-        let path = Path::new("/a/b/c/../../d");
-        assert_eq!(normalize_path(path), PathBuf::from("/a/d"));
-    }
-
-    #[test]
-    fn normalize_path_preserves_root_on_excessive_parents() {
-        let path = Path::new("/a/../../../b");
-        assert_eq!(normalize_path(path), PathBuf::from("/b"));
-    }
-
-    #[test]
-    fn parse_mount_spec_resolved_resolves_relative_src_and_dst() {
-        let cwd = std::env::current_dir().unwrap();
-        let mount = parse_mount_spec_resolved("my-project").unwrap();
-        let expected = cwd.join("my-project").display().to_string();
-
-        assert_eq!(mount.src, expected);
-        assert_eq!(mount.dst, expected);
-        assert!(!mount.readonly);
-    }
-
-    #[test]
-    fn parse_mount_spec_resolved_resolves_relative_src_with_explicit_dst() {
-        let cwd = std::env::current_dir().unwrap();
-        let mount = parse_mount_spec_resolved("my-project:/workspace/project").unwrap();
-
-        assert_eq!(mount.src, cwd.join("my-project").display().to_string());
-        assert_eq!(mount.dst, "/workspace/project");
-        assert!(!mount.readonly);
-    }
-
-    #[test]
-    fn parse_mount_spec_resolved_normalizes_dotdot_in_relative_path() {
-        let cwd = std::env::current_dir().unwrap();
-        let mount = parse_mount_spec_resolved("../sibling-project").unwrap();
-        let expected = cwd.parent().unwrap().join("sibling-project");
-
-        assert_eq!(mount.src, expected.display().to_string());
-        assert_eq!(mount.dst, expected.display().to_string());
-        assert!(!mount.src.contains(".."));
-    }
-
-    #[test]
-    fn parse_mount_spec_resolved_normalizes_dot_path() {
-        let cwd = std::env::current_dir().unwrap();
-        let mount = parse_mount_spec_resolved(".").unwrap();
-
-        assert_eq!(mount.src, cwd.display().to_string());
-        assert_eq!(mount.dst, cwd.display().to_string());
-    }
-
-    #[test]
-    fn parse_mount_spec_does_not_resolve_relative_paths() {
-        let mount = parse_mount_spec("my-project").unwrap();
-
-        assert_eq!(mount.src, "my-project");
-        assert_eq!(mount.dst, "my-project");
-    }
-
-    #[test]
-    fn current_dir_workspace_uses_same_host_and_container_path() {
-        let dir = tempdir().unwrap();
-        let workspace = current_dir_workspace(dir.path()).unwrap();
-
-        assert_eq!(
-            workspace.workdir,
-            dir.path().canonicalize().unwrap().display().to_string()
-        );
-        assert_eq!(workspace.mounts.len(), 1);
-        assert_eq!(workspace.mounts[0].src, workspace.mounts[0].dst);
-    }
-
-    #[test]
-    fn saved_workspace_match_depth_matches_host_workdir_parent_of_mounts() {
-        let temp = tempdir().unwrap();
-        let workspace_root = temp.path().join("monorepo");
-        let repo_a = workspace_root.join("jackin");
-        let repo_b = workspace_root.join("jackin-dev");
-        std::fs::create_dir_all(&repo_a).unwrap();
-        std::fs::create_dir_all(&repo_b).unwrap();
-
-        let canonical_root = workspace_root.canonicalize().unwrap();
-        let workspace = WorkspaceConfig {
-            workdir: canonical_root.display().to_string(),
-            mounts: vec![
-                MountConfig {
-                    src: repo_a.canonicalize().unwrap().display().to_string(),
-                    dst: "/workspace/jackin".to_string(),
-                    readonly: false,
-                },
-                MountConfig {
-                    src: repo_b.canonicalize().unwrap().display().to_string(),
-                    dst: "/workspace/jackin-dev".to_string(),
-                    readonly: false,
-                },
-            ],
-            allowed_agents: vec![],
-            default_agent: None,
-            last_agent: None,
-        };
-
-        assert_eq!(
-            saved_workspace_match_depth(&workspace, &canonical_root),
-            Some(canonical_root.components().count())
-        );
-    }
-
-    #[test]
-    fn saved_workspace_match_depth_still_matches_nested_path_under_mount_root() {
-        let temp = tempdir().unwrap();
-        let project_dir = temp.path().join("project");
-        let nested_dir = project_dir.join("src/bin");
-        std::fs::create_dir_all(&nested_dir).unwrap();
-
-        let workspace = WorkspaceConfig {
-            workdir: "/workspace".to_string(),
-            mounts: vec![MountConfig {
-                src: project_dir.canonicalize().unwrap().display().to_string(),
-                dst: "/workspace".to_string(),
-                readonly: false,
-            }],
-            allowed_agents: vec![],
-            default_agent: None,
-            last_agent: None,
-        };
-
-        assert_eq!(
-            saved_workspace_match_depth(&workspace, &nested_dir),
-            Some(project_dir.canonicalize().unwrap().components().count())
-        );
-    }
-
-    #[test]
-    fn resolves_saved_workspace_and_rejects_disallowed_agent() {
-        let mut config = crate::config::AppConfig::default();
-        config.agents.insert(
-            "agent-smith".to_string(),
-            crate::config::AgentSource {
-                git: "https://github.com/jackin-project/jackin-agent-smith.git".to_string(),
-                trusted: true,
-                claude: None,
-            },
-        );
-        config.workspaces.insert(
-            "big-monorepo".to_string(),
-            WorkspaceConfig {
-                workdir: "/workspace/project".to_string(),
-                mounts: vec![MountConfig {
-                    src: "/tmp/project".to_string(),
-                    dst: "/workspace/project".to_string(),
-                    readonly: false,
-                }],
-                allowed_agents: vec!["agent-smith".to_string()],
-                default_agent: Some("agent-smith".to_string()),
-                last_agent: None,
-            },
-        );
-
-        let cwd = std::env::temp_dir();
-        let error = resolve_load_workspace(
-            &config,
-            &crate::selector::ClassSelector::new(None, "neo"),
-            &cwd,
-            LoadWorkspaceInput::Saved("big-monorepo".to_string()),
-            &[],
-        )
-        .unwrap_err();
-
-        assert!(error.to_string().contains("is not allowed by workspace"));
-    }
-
-    #[test]
-    fn saved_workspace_label_uses_workspace_name() {
-        let temp = tempdir().unwrap();
-        let mount_src = temp.path().join("project");
-        std::fs::create_dir_all(&mount_src).unwrap();
-
-        let mut config = crate::config::AppConfig::default();
-        config.agents.insert(
-            "agent-smith".to_string(),
-            crate::config::AgentSource {
-                git: "https://github.com/jackin-project/jackin-agent-smith.git".to_string(),
-                trusted: true,
-                claude: None,
-            },
-        );
-        config.workspaces.insert(
-            "big-monorepo".to_string(),
-            WorkspaceConfig {
-                workdir: "/workspace/project".to_string(),
-                mounts: vec![MountConfig {
-                    src: mount_src.display().to_string(),
-                    dst: "/workspace/project".to_string(),
-                    readonly: false,
-                }],
-                allowed_agents: vec![],
-                default_agent: None,
-                last_agent: None,
-            },
-        );
-
-        let cwd = std::env::temp_dir();
-        let resolved = resolve_load_workspace(
-            &config,
-            &crate::selector::ClassSelector::new(None, "agent-smith"),
-            &cwd,
-            LoadWorkspaceInput::Saved("big-monorepo".to_string()),
-            &[],
-        )
-        .unwrap();
-
-        assert_eq!(resolved.label, "big-monorepo");
-        assert_eq!(resolved.workdir, "/workspace/project");
-    }
-
-    #[test]
-    fn resolves_same_path_relative_target_to_absolute_workdir() {
-        let temp = tempdir().unwrap();
-        let project_dir = temp.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        // The cwd parameter is used to resolve relative paths — no need
-        // to mutate the global process CWD.
-        let resolved = resolve_load_workspace(
-            &crate::config::AppConfig::default(),
-            &crate::selector::ClassSelector::new(None, "agent-smith"),
-            temp.path(),
-            LoadWorkspaceInput::Path {
-                src: "./project".to_string(),
-                dst: "./project".to_string(),
-            },
-            &[],
-        )
-        .unwrap();
-
-        let expected = project_dir.canonicalize().unwrap().display().to_string();
-        assert_eq!(resolved.workdir, expected);
-        assert_eq!(resolved.mounts[0].dst, expected);
-    }
-
-    #[test]
-    fn resolves_global_mounts_with_tilde_sources() {
-        let home = std::env::var("HOME").unwrap();
-        let cwd = tempdir().unwrap();
-        let mut config = crate::config::AppConfig::default();
-        config.add_mount(
-            "home",
-            MountConfig {
-                src: "~".to_string(),
-                dst: "/home/claude/home".to_string(),
-                readonly: true,
-            },
-            None,
-        );
-
-        let resolved = resolve_load_workspace(
-            &config,
-            &crate::selector::ClassSelector::new(None, "agent-smith"),
-            cwd.path(),
-            LoadWorkspaceInput::CurrentDir,
-            &[],
-        )
-        .unwrap();
-
-        assert!(
-            resolved
-                .mounts
-                .iter()
-                .any(|mount| mount.dst == "/home/claude/home"
-                    && mount.src == home
-                    && mount.readonly)
-        );
-    }
-
-    #[test]
-    fn resolve_with_ad_hoc_mounts_merges_correctly() {
-        let temp = tempdir().unwrap();
-        let mount_src = temp.path().join("project");
-        let extra_src = temp.path().join("extra");
-        std::fs::create_dir_all(&mount_src).unwrap();
-        std::fs::create_dir_all(&extra_src).unwrap();
-
-        let mut config = crate::config::AppConfig::default();
-        config.agents.insert(
-            "agent-smith".to_string(),
-            crate::config::AgentSource {
-                git: "https://github.com/jackin-project/jackin-agent-smith.git".to_string(),
-                trusted: true,
-                claude: None,
-            },
-        );
-        config.workspaces.insert(
-            "my-ws".to_string(),
-            WorkspaceConfig {
-                workdir: "/workspace/project".to_string(),
-                mounts: vec![MountConfig {
-                    src: mount_src.display().to_string(),
-                    dst: "/workspace/project".to_string(),
-                    readonly: false,
-                }],
-                allowed_agents: vec![],
-                default_agent: None,
-                last_agent: None,
-            },
-        );
-
-        let cwd = std::env::temp_dir();
-        let resolved = resolve_load_workspace(
-            &config,
-            &crate::selector::ClassSelector::new(None, "agent-smith"),
-            &cwd,
-            LoadWorkspaceInput::Saved("my-ws".to_string()),
-            &[MountConfig {
-                src: extra_src.display().to_string(),
-                dst: "/extra".to_string(),
-                readonly: true,
-            }],
-        )
-        .unwrap();
-
-        assert_eq!(resolved.mounts.len(), 2);
-        assert!(
-            resolved
-                .mounts
-                .iter()
-                .any(|m| m.dst == "/extra" && m.readonly)
-        );
-    }
-
-    #[test]
-    fn resolve_with_ad_hoc_mount_dst_conflict_errors() {
-        let temp = tempdir().unwrap();
-        let mount_src = temp.path().join("project");
-        std::fs::create_dir_all(&mount_src).unwrap();
-
-        let mut config = crate::config::AppConfig::default();
-        config.workspaces.insert(
-            "my-ws".to_string(),
-            WorkspaceConfig {
-                workdir: "/workspace/project".to_string(),
-                mounts: vec![MountConfig {
-                    src: mount_src.display().to_string(),
-                    dst: "/workspace/project".to_string(),
-                    readonly: false,
-                }],
-                allowed_agents: vec![],
-                default_agent: None,
-                last_agent: None,
-            },
-        );
-
-        let cwd = std::env::temp_dir();
-        let error = resolve_load_workspace(
-            &config,
-            &crate::selector::ClassSelector::new(None, "agent-smith"),
-            &cwd,
-            LoadWorkspaceInput::Saved("my-ws".to_string()),
-            &[MountConfig {
-                src: mount_src.display().to_string(),
-                dst: "/workspace/project".to_string(),
-                readonly: false,
-            }],
-        )
-        .unwrap_err();
-
-        assert!(
-            error
-                .to_string()
-                .contains("ad-hoc mount destination conflicts")
-        );
-    }
 
     // -- validate_workspace_config: workdir vs mount destination coverage ------
 
@@ -1131,421 +205,5 @@ mod tests {
             last_agent: None,
         };
         validate_workspace_config("test", &ws).unwrap();
-    }
-
-    // -- find_sensitive_mounts ------------------------------------------------
-
-    fn mount(src: &str) -> MountConfig {
-        MountConfig {
-            src: src.to_string(),
-            dst: "/container/path".to_string(),
-            readonly: false,
-        }
-    }
-
-    #[test]
-    fn detects_ssh_mount() {
-        let mounts = vec![mount("/home/user/.ssh")];
-        let hits = find_sensitive_mounts(&mounts);
-        assert_eq!(hits.len(), 1);
-        assert_eq!(hits[0].src, "/home/user/.ssh");
-        assert!(hits[0].reason.contains("SSH"));
-    }
-
-    #[test]
-    fn detects_aws_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.aws")]);
-        assert_eq!(hits.len(), 1);
-        assert!(hits[0].reason.contains("AWS"));
-    }
-
-    #[test]
-    fn detects_gnupg_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.gnupg")]);
-        assert_eq!(hits.len(), 1);
-        assert!(hits[0].reason.contains("GPG"));
-    }
-
-    #[test]
-    fn detects_gcloud_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.config/gcloud")]);
-        assert_eq!(hits.len(), 1);
-        assert!(hits[0].reason.contains("Google Cloud"));
-    }
-
-    #[test]
-    fn detects_kube_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.kube")]);
-        assert_eq!(hits.len(), 1);
-        assert!(hits[0].reason.contains("Kubernetes"));
-    }
-
-    #[test]
-    fn detects_docker_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.docker")]);
-        assert_eq!(hits.len(), 1);
-        assert!(hits[0].reason.contains("Docker"));
-    }
-
-    #[test]
-    fn ignores_safe_mounts() {
-        let mounts = vec![
-            mount("/home/user/projects"),
-            mount("/tmp/workspace"),
-            mount("/var/data"),
-        ];
-        assert!(find_sensitive_mounts(&mounts).is_empty());
-    }
-
-    #[test]
-    fn detects_multiple_sensitive_mounts() {
-        let mounts = vec![
-            mount("/home/user/.ssh"),
-            mount("/home/user/projects"),
-            mount("/home/user/.aws"),
-        ];
-        let hits = find_sensitive_mounts(&mounts);
-        assert_eq!(hits.len(), 2);
-    }
-
-    #[test]
-    fn handles_trailing_slash_on_sensitive_mount() {
-        let hits = find_sensitive_mounts(&[mount("/home/user/.ssh/")]);
-        assert_eq!(hits.len(), 1);
-    }
-
-    #[test]
-    fn does_not_match_partial_name() {
-        // ".sshd" should NOT match ".ssh"
-        let hits = find_sensitive_mounts(&[mount("/home/user/.sshd")]);
-        assert!(hits.is_empty());
-    }
-
-    #[test]
-    fn covers_is_false_for_equal_mounts() {
-        let a = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let b = a.clone();
-        assert!(!covers(&a, &b));
-    }
-
-    #[test]
-    fn covers_is_true_for_exact_ancestor_with_matching_suffix() {
-        let parent = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/a/b".into(),
-            dst: "/a/b".into(),
-            readonly: false,
-        };
-        assert!(covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_is_true_for_deep_ancestor_with_matching_suffix() {
-        let parent = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/a/b/c/d".into(),
-            dst: "/a/b/c/d".into(),
-            readonly: false,
-        };
-        assert!(covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_is_true_when_src_and_dst_differ_but_offsets_match() {
-        let parent = MountConfig {
-            src: "/host/root".into(),
-            dst: "/container/root".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/host/root/sub".into(),
-            dst: "/container/root/sub".into(),
-            readonly: false,
-        };
-        assert!(covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_is_false_when_src_nests_but_dst_offsets_differ() {
-        let parent = MountConfig {
-            src: "/host/root".into(),
-            dst: "/container/a".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/host/root/sub".into(),
-            dst: "/container/b/sub".into(),
-            readonly: false,
-        };
-        assert!(!covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_is_false_when_src_does_not_nest() {
-        let a = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let b = MountConfig {
-            src: "/b".into(),
-            dst: "/b".into(),
-            readonly: false,
-        };
-        assert!(!covers(&a, &b));
-    }
-
-    #[test]
-    fn covers_is_false_for_sibling_prefix_match() {
-        // `/a-x` is not a child of `/a`, even though "/a-x".starts_with("/a").
-        let parent = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/a-x".into(),
-            dst: "/a-x".into(),
-            readonly: false,
-        };
-        assert!(!covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_normalizes_trailing_slashes() {
-        let parent = MountConfig {
-            src: "/a/".into(),
-            dst: "/a/".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/a/b".into(),
-            dst: "/a/b".into(),
-            readonly: false,
-        };
-        assert!(covers(&parent, &child));
-    }
-
-    #[test]
-    fn covers_handles_different_readonly_flags() {
-        // `covers` is purely path-based. Readonly mismatches are caught by plan_collapse.
-        let parent = MountConfig {
-            src: "/a".into(),
-            dst: "/a".into(),
-            readonly: false,
-        };
-        let child = MountConfig {
-            src: "/a/b".into(),
-            dst: "/a/b".into(),
-            readonly: true,
-        };
-        assert!(covers(&parent, &child));
-    }
-
-    fn mk(src: &str, dst: &str, ro: bool) -> MountConfig {
-        MountConfig {
-            src: src.into(),
-            dst: dst.into(),
-            readonly: ro,
-        }
-    }
-
-    #[test]
-    fn plan_collapse_empty_input_returns_empty_plan() {
-        let plan = plan_collapse(&[], &[]).unwrap();
-        assert!(plan.kept.is_empty());
-        assert!(plan.removed.is_empty());
-    }
-
-    #[test]
-    fn plan_collapse_preserves_unrelated_mounts() {
-        let mounts = vec![mk("/a", "/a", false), mk("/b", "/b", false)];
-        let plan = plan_collapse(&mounts, &[0, 1]).unwrap();
-        assert_eq!(plan.kept, mounts);
-        assert!(plan.removed.is_empty());
-    }
-
-    #[test]
-    fn plan_collapse_collapses_single_child_under_new_parent() {
-        let mounts = vec![
-            mk("/a/b", "/a/b", false), // pre-existing child (index 0)
-            mk("/a", "/a", false),     // new parent (index 1)
-        ];
-        let plan = plan_collapse(&mounts, &[1]).unwrap();
-        assert_eq!(plan.kept, vec![mk("/a", "/a", false)]);
-        assert_eq!(plan.removed.len(), 1);
-        assert_eq!(plan.removed[0].child, mk("/a/b", "/a/b", false));
-        assert_eq!(plan.removed[0].covered_by, mk("/a", "/a", false));
-    }
-
-    #[test]
-    fn plan_collapse_collapses_multiple_children_under_new_parent() {
-        let mounts = vec![
-            mk("/a/b", "/a/b", false),
-            mk("/a/c", "/a/c", false),
-            mk("/a", "/a", false),
-        ];
-        let plan = plan_collapse(&mounts, &[2]).unwrap();
-        assert_eq!(plan.kept, vec![mk("/a", "/a", false)]);
-        assert_eq!(plan.removed.len(), 2);
-    }
-
-    #[test]
-    fn plan_collapse_handles_transitive_chain() {
-        // A ⊃ B ⊃ C, all present in input; B and C both find A as a parent.
-        let mounts = vec![
-            mk("/a", "/a", false),
-            mk("/a/b", "/a/b", false),
-            mk("/a/b/c", "/a/b/c", false),
-        ];
-        let plan = plan_collapse(&mounts, &[0, 1, 2]).unwrap();
-        assert_eq!(plan.kept, vec![mk("/a", "/a", false)]);
-        assert_eq!(plan.removed.len(), 2);
-        // Both B and C are covered by A.
-        for rem in &plan.removed {
-            assert_eq!(rem.covered_by, mk("/a", "/a", false));
-        }
-    }
-
-    #[test]
-    fn plan_collapse_flags_pre_existing_violation_as_removal_when_nothing_new() {
-        // Neither index is in new_indexes — both pre-existing. Library returns
-        // Ok with removal; CLI will inspect origin and decide to reject or
-        // proceed with --prune.
-        let mounts = vec![mk("/a", "/a", false), mk("/a/b", "/a/b", false)];
-        let plan = plan_collapse(&mounts, &[]).unwrap();
-        assert_eq!(plan.kept, vec![mk("/a", "/a", false)]);
-        assert_eq!(plan.removed.len(), 1);
-    }
-
-    #[test]
-    fn plan_collapse_errors_on_readonly_mismatch_rw_parent_ro_child() {
-        let mounts = vec![
-            mk("/a/b", "/a/b", true), // ro child
-            mk("/a", "/a", false),    // rw parent (new)
-        ];
-        let err = plan_collapse(&mounts, &[1]).unwrap_err();
-        match err {
-            CollapseError::ReadonlyMismatch { parent, child } => {
-                assert_eq!(parent, mk("/a", "/a", false));
-                assert_eq!(child, mk("/a/b", "/a/b", true));
-            }
-            other => panic!("expected ReadonlyMismatch, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn plan_collapse_errors_on_readonly_mismatch_ro_parent_rw_child() {
-        let mounts = vec![
-            mk("/a/b", "/a/b", false), // rw child
-            mk("/a", "/a", true),      // ro parent (new)
-        ];
-        let err = plan_collapse(&mounts, &[1]).unwrap_err();
-        assert!(matches!(err, CollapseError::ReadonlyMismatch { .. }));
-    }
-
-    #[test]
-    fn plan_collapse_errors_on_new_child_under_existing_parent() {
-        // Parent at index 0 is pre-existing. Child at index 1 is new.
-        let mounts = vec![
-            mk("/a", "/a", false),     // existing parent
-            mk("/a/b", "/a/b", false), // new child
-        ];
-        let err = plan_collapse(&mounts, &[1]).unwrap_err();
-        match err {
-            CollapseError::ChildUnderExistingParent { parent, child } => {
-                assert_eq!(parent, mk("/a", "/a", false));
-                assert_eq!(child, mk("/a/b", "/a/b", false));
-            }
-            other => panic!("expected ChildUnderExistingParent, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn plan_collapse_allows_new_child_when_new_parent_is_also_in_same_edit() {
-        // Both parent and child introduced in the same edit — not a "child
-        // under existing parent" case; child is just redundant and gets
-        // collapsed normally.
-        let mounts = vec![mk("/a/b", "/a/b", false), mk("/a", "/a", false)];
-        let plan = plan_collapse(&mounts, &[0, 1]).unwrap();
-        assert_eq!(plan.kept, vec![mk("/a", "/a", false)]);
-        assert_eq!(plan.removed.len(), 1);
-    }
-
-    #[test]
-    fn plan_collapse_error_message_mentions_both_paths() {
-        let mounts = vec![mk("/a/b", "/a/b", true), mk("/a", "/a", false)];
-        let err = plan_collapse(&mounts, &[1]).unwrap_err();
-        let msg = err.to_string();
-        assert!(msg.contains("/a"));
-        assert!(msg.contains("/a/b"));
-        assert!(msg.contains("readonly"));
-    }
-
-    /// Exhaustive check: for any plan produced on a valid (no-error) input,
-    /// re-planning on `plan.kept` must be a no-op.
-    #[test]
-    fn plan_collapse_is_idempotent() {
-        let inputs: Vec<Vec<MountConfig>> = vec![
-            vec![],
-            vec![mk("/a", "/a", false)],
-            vec![mk("/a", "/a", false), mk("/b", "/b", false)],
-            vec![mk("/a", "/a", false), mk("/a/b", "/a/b", false)],
-            vec![
-                mk("/a", "/a", false),
-                mk("/a/b", "/a/b", false),
-                mk("/a/b/c", "/a/b/c", false),
-                mk("/x", "/x", true),
-            ],
-        ];
-        for input in inputs {
-            let indexes: Vec<usize> = (0..input.len()).collect();
-            let plan = plan_collapse(&input, &indexes).unwrap();
-            let second = plan_collapse(&plan.kept, &[]).unwrap();
-            assert!(
-                second.removed.is_empty(),
-                "plan.kept should be rule-C compliant, but re-plan removed {} entries",
-                second.removed.len(),
-            );
-            assert_eq!(second.kept, plan.kept);
-        }
-    }
-
-    #[test]
-    fn plan_collapse_result_satisfies_invariant() {
-        // After planning, no pair in `kept` covers another pair in `kept`.
-        let mounts = vec![
-            mk("/a", "/a", false),
-            mk("/a/b", "/a/b", false),
-            mk("/a/c/d", "/a/c/d", false),
-            mk("/x/y", "/x/y", true),
-            mk("/x", "/x", true),
-        ];
-        let indexes: Vec<usize> = (0..mounts.len()).collect();
-        let plan = plan_collapse(&mounts, &indexes).unwrap();
-        for (i, a) in plan.kept.iter().enumerate() {
-            for (j, b) in plan.kept.iter().enumerate() {
-                if i != j {
-                    assert!(
-                        !covers(a, b),
-                        "invariant violated: {:?} covers {:?} in kept set",
-                        a,
-                        b,
-                    );
-                }
-            }
-        }
     }
 }
